requirements.txt
# 项目依赖

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 机器学习
scikit-learn>=1.0.0

# 量化分析
backtrader>=1.9.76

# HTTP服务
flask>=2.0.0

# 缓存
redis>=4.0.0

# 配置
pyyaml>=6.0

# HTTP请求
requests>=2.25.0

# 异步处理
asyncio
aiohttp>=3.8.0

# 数据可视化
matplotlib>=3.5.0